# Core/StickyBoard.Core/Docs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "StickyBoard API",
    "description": "REST API for StickyBoard cross-platform workspace (Academic Project)",
    "contact": {
      "name": "Alexandre Emond",
      "url": "https://aedev.pro"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/cards/{cardId}/comments": {
      "get": {
        "tags": [
          "CardComment"
        ],
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Invalid input",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "AUTH_INVALID",
                  "message": "Authentication required or invalid token",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "User does not have access to this resource",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Requested resource does not exist",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "SERVER_ERROR",
                  "message": "Unexpected server error",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CardComment"
        ],
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCommentCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCommentCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CardCommentCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Invalid input",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "AUTH_INVALID",
                  "message": "Authentication required or invalid token",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "User does not have access to this resource",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Requested resource does not exist",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "SERVER_ERROR",
                  "message": "Unexpected server error",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/{id}": {
      "get": {
        "tags": [
          "Cards"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Invalid input",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "AUTH_INVALID",
                  "message": "Authentication required or invalid token",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "User does not have access to this resource",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Requested resource does not exist",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "SERVER_ERROR",
                  "message": "Unexpected server error",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cards"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CardUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CardUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Invalid input",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "AUTH_INVALID",
                  "message": "Authentication required or invalid token",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "User does not have access to this resource",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Requested resource does not exist",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "SERVER_ERROR",
                  "message": "Unexpected server error",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Cards"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Invalid input",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "AUTH_INVALID",
                  "message": "Authentication required or invalid token",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "User does not have access to this resource",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Requested resource does not exist",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "example": {
                  "code": "SERVER_ERROR",
                  "message": "Unexpected server error",
                  "details": "Additional debug info in development mode only"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardCommentCreateDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CardStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "CardUpdateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/CardStatus"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "assigneeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "sectionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "tabId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ErrorCode": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization (use: Bearer <token>)",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKey": {
        "type": "apiKey",
        "description": "Worker API Key (use: ApiKey <your_key>)",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [],
      "ApiKey": []
    }
  ]
}